4.21: bloglist expansion step9

+++ controllers/blogs.js
@@ -35,6 +35,20 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  console.log('-- decodedToken.id: ', decodedToken.id)
+  console.log('-- blog:', blog)
+
+  if (decodedToken.id.toString() !== blog.user.toString()) {
+    return response.status(400).json({ error: 'invalid user' })
+  }
+  
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
